DZ_TEST/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
pip==22.3.1
pluggy==1.0.0
pytest==7.3.1
python-dotenv==1.0.0
requests==2.29.0
setuptools==65.5.1
tomli==2.0.1
urllib3==1.26.15
wheel==0.38.4